refactor(ios): extract Clerk network middleware from JovieApp
Behavior-preserving move (JovieApp.swift 458 -> 183 lines): the Clerk
native-ticket sign-in plumbing (request/response middlewares,
NativeClerkDeviceTokenFallback actor, Keychain-backed token store,
response parser, DEBUG NativeTicketSignInDiagnostics, and the
makeJovieClerkOptions factory) moves verbatim to
App/ClerkNetworkMiddleware.swift. JovieApp.swift keeps @main, the launch
configuration resolver, and JovieAppDelegate. pbxproj wired with
explicit entries.
